fluent-langneg: Make negotiateLanguages a named export
